lodash added parsers are refactored UMD loader added
