+++ lib/parser.js
@@ -6,5 +6,5 @@
 
 module.exports = {
     RedParser: require('./parser/RedParser.js'),
-    PaginateableParser: require('./parser/PaginateableParser.js')
+    PaginateableParser: require('./parser/PaginatableParser.js')
 };
