+++ lib/parser/debug.js
@@ -2,7 +2,7 @@ var debug = function () {
     if (
         !(window.DEBUG instanceof Array) ||
         (
-            window.DEBUG.indexOf('PaginateableParser') === -1 &&
+            window.DEBUG.indexOf('PaginatableParser') === -1 &&
             window.DEBUG.indexOf('RedParser') === -1
         )
     ) {
