+++ lodash.md
@@ -1,0 +1,11 @@
+### lodash
+
+Lodash installed automatically after `npm install` to `lodash.js` file
+
+#### adding lodash modules
+
+1. open `install-lodash.sh`
+2. add modules to `include_modules` variable
+3. run `npm run lodash`
+
+other developers will give update after `npm install`
