+++ package.json
@@ -28,12 +28,16 @@
   },
   "main": "./lib/parser",
   "scripts": {
-    "test": "mocha-phantomjs -p `which phantomjs` tests/redparser.html"
+    "test": "mocha-phantomjs -p `which phantomjs` tests/redparser.html",
+    "lodash": "./install-lodash.sh --update",
+    "postinstall": "npm run lodash"
   },
   "engines": {
     "node": ">= 0.4.0"
   },
-  "dependencies": {},
+  "dependencies": {
+    "lodash-cli": "^3.10.1"
+  },
   "devDependencies": {
     "chai": "^3.3.0",
     "mocha": "^2.3.3",
