+++ tests/redparser.html
@@ -126,8 +126,9 @@
     });
 </script>
 
+<script src="../lib/lodash.js"></script>
 <script src="../lib/parser/RedParser.js"></script>
-<script src="../lib/parser/PaginateableParser.js"></script>
+<script src="../lib/parser/PaginatableParser.js"></script>
 <script src="redparser_test.js"></script>
 <script>
     mocha.checkLeaks();
